refactor: move ldtk registrations and bundles out of sokoban module and into new wall module
