+++ src/graveyard/mod.rs
@@ -8,6 +8,7 @@ pub mod goal;
 pub mod gravestone;
 pub mod movement_table;
 pub mod sokoban;
+pub mod wall;
 pub mod willo;
 pub mod wind;
